requirements.txt
panel
pandas
python-docx
odfpy
